requirements.txt
joblib==1.3.2
networkx==3.2.1
numpy==1.26.4
ogb==1.3.6
pandas==2.2.2
PyYAML==6.0.1
PyYAML==6.0.1
rdkit==2023.9.5
scikit_learn==1.4.1.post1
scipy==1.13.0
torch==2.1.0+cu118
torch_cluster==1.6.3+pt21cu118
torch_geometric==2.5.3
tqdm==4.66.2
